Refactored the integration test to the normal test tree to fix missing Sonar coverage
