+++ pom.xml
@@ -302,7 +302,7 @@
 					</execution>
 					<execution>
 						<id>post-unit-test</id>
-						<phase>verify</phase>
+						<phase>test</phase>
 						<goals>
 							<goal>report</goal>
 						</goals>
@@ -315,10 +315,9 @@
 						<id>pre-integration-test</id>
 						<phase>pre-integration-test</phase>
 						<goals>
-							<goal>prepare-agent-integration</goal>
+							<goal>prepare-agent</goal>
 						</goals>
 						<configuration>
-							<dumpOnExit>true</dumpOnExit>
 							<!-- Sets the path to the file which contains the execution data. -->
 							<destFile>${sonar.jacoco.reports}/jacoco-it.exec</destFile>
 							<propertyName>failsafeArgLine</propertyName>
@@ -328,9 +327,11 @@
 						<id>post-integration-test</id>
 						<phase>post-integration-test</phase>
 						<goals>
-							<goal>report-aggregate</goal>
+							<goal>report</goal>
 						</goals>
 						<configuration>
+							<!-- Sets the path to the file which contains the execution data. -->
+							<dataFile>${sonar.jacoco.reports}/jacoco-it.exec</dataFile>
 							<!-- Sets the output directory for the code coverage report. -->
 							<outputDirectory>${sonar.jacoco.reports}/jacoco-it</outputDirectory>
 						</configuration>
